fix layout issues and add download options component
