+++ client/blocks/get-apps/desktop-download-card.tsx
@@ -1,36 +1,12 @@
-import { useTranslate } from 'i18n-calypso';
 import { useMemo } from 'react';
-import PopoverMenuItem from 'calypso/components/popover-menu/item';
-import SplitButton from 'calypso/components/split-button';
-import SVGIcon from 'calypso/components/svg-icon';
 import userAgent from 'calypso/lib/user-agent';
 import { AppsCard } from './apps-card';
 import { PlatformType, type DesktopAppConfig } from './apps-config';
+import { DesktopDownloadOptions } from './desktop-download-options';
 
-interface AlsoAvailableConfig {
-	name: string;
-	icon: string;
-	iconName: string;
-	link: string;
-	onClick: () => void;
-}
-
-const AlsoAvailable: React.FC< { config: AlsoAvailableConfig } > = ( { config } ) => (
-	<a href={ config.link } onClick={ config.onClick } className="get-apps__desktop-link">
-		<SVGIcon
-			classes=""
-			aria-hidden="true"
-			name={ config.iconName }
-			size={ 16 }
-			icon={ config.icon }
-		/>
-		{ config.name }
-	</a>
-);
-
-interface DesktopDownloadCardProps {
+type DesktopDownloadCardProps = {
 	appConfig: DesktopAppConfig;
-}
+};
 
 const getCurrentPlatform = (): PlatformType => {
 	const platformName = navigator.platform;
@@ -49,7 +25,6 @@ const getCurrentPlatform = (): PlatformType => {
 };
 
 const DesktopDownloadCard: React.FC< DesktopDownloadCardProps > = ( { appConfig } ) => {
-	const translate = useTranslate();
 	const { isMobile } = userAgent;
 	const platform = useMemo( () => getCurrentPlatform(), [] );
 
@@ -58,66 +33,6 @@ const DesktopDownloadCard: React.FC< DesktopDownloadCardProps > = ( { appConfig 
 		[ appConfig.platforms, platform ]
 	);
 
-	const getDesktopDeviceDownloadOptions = () => {
-		if ( ! currentPlatformConfig ) {
-			return (
-				<div className="get-apps__not-available">
-					{ translate( 'Not available for your platform. Available for:' ) }
-					<div className="get-apps__also-available-list">
-						{ Object.entries( appConfig.platforms ).map( ( [ key, config ] ) => (
-							<AlsoAvailable key={ key } config={ config } />
-						) ) }
-					</div>
-				</div>
-			);
-		}
-
-		return (
-			<>
-				<div className="get-apps__desktop-button">
-					<SplitButton
-						whiteSeparator={ appConfig.isPrimary }
-						primary={ appConfig.isPrimary }
-						label={ currentPlatformConfig.buttonText }
-						icon={
-							<SVGIcon
-								classes="get-apps__desktop-button-icon"
-								aria-hidden="true"
-								name={ currentPlatformConfig.iconName }
-								size={ 16 }
-								icon={ currentPlatformConfig.icon }
-							/>
-						}
-						onClick={ currentPlatformConfig.onClick }
-						href={ currentPlatformConfig.link }
-					>
-						{ Object.entries( appConfig.platforms )
-							.filter( ( [ , config ] ) => config.group === currentPlatformConfig.group )
-							.map( ( [ key, config ] ) => (
-								<PopoverMenuItem key={ key } href={ config.link } onClick={ config.onClick }>
-									{ config.name }
-								</PopoverMenuItem>
-							) ) }
-					</SplitButton>
-				</div>
-
-				<div className="get-apps__also-available">
-					<div className="get-apps__also-available-title">
-						{ translate( 'Also available for:' ) }
-					</div>
-
-					<div className="get-apps__also-available-list">
-						{ Object.entries( appConfig.platforms )
-							.filter( ( [ , config ] ) => config.group !== currentPlatformConfig.group )
-							.map( ( [ key, config ] ) => (
-								<AlsoAvailable key={ key } config={ config } />
-							) ) }
-					</div>
-				</div>
-			</>
-		);
-	};
-
 	return (
 		<AppsCard
 			logo={ appConfig.logo }
@@ -125,17 +40,11 @@ const DesktopDownloadCard: React.FC< DesktopDownloadCardProps > = ( { appConfig 
 			title={ appConfig.title }
 			subtitle={ appConfig.subtitle }
 		>
-			{ isMobile ? (
-				<div className="get-apps__desktop-link">
-					{ translate( 'Visit {{a}}desktop.wordpress.com{{/a}} on your desktop.', {
-						components: {
-							a: <a href="https://desktop.wordpress.com" />,
-						},
-					} ) }
-				</div>
-			) : (
-				getDesktopDeviceDownloadOptions()
-			) }
+			<DesktopDownloadOptions
+				appConfig={ appConfig }
+				currentPlatformConfig={ currentPlatformConfig }
+				isMobile={ isMobile }
+			/>
 		</AppsCard>
 	);
 };
